[FIX] hw_drivers: force touch in kiosk mode
Before this commit, when Chromium was running in kiosk mode under
Wayland, it would not detect that the input device was a touchscreen.
This resulted in touchscreen-specific code (such as hiding the cursor in
the self-order kiosk) not working.

After this commit, we pass the `--touch-events` flag to Chromium when
running in kiosk mode, which forces touch events to be enabled.

closes odoo/odoo#206025

Signed-off-by: Louis Travaux (lotr) <lotr@odoo.com>