[FIX] website: fix hidden snippet forced visibility
Steps to Reproduce :

- Drag and drop a "Text" snippet.
- Click on "Hide on desktop".
- Toggle on the eye button to force its visibility.
- Click on a text outside the snippet & press enter.
- Bug: the invisible snippet disappears.

The issue comes from the fact that when we press Enter, a new paragraph
is created and a history rollback is triggered at that moment. So when
we go back in the history because of that rollback, the invisible
snippet loses the "o_snippet_override_invisible" class that had just
been applied.

Actually, adding and removing this class should not be inside the
observer. This commit prevents that from happening.

task-4690318

closes odoo/odoo#206527

X-original-commit: c7bdf32c34f72dfa82ac6c52e018fe76acc8e012
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>
Signed-off-by: Benjamin Vray (bvr) <bvr@odoo.com>