[FIX] point_of_sale: restore customer display QR code
Previously, when using a payment method integrated with the Bank App (QR
code), a QR code was shown on the customer display.

This functionality was lost during the refactoring of the customer
display into a standalone OWL app:
https://github.com/odoo/odoo/commit/acf78c27b12cf014cd80d20ea5853e63ab9ca03f

It was later removed entirely with the deletion of the
point_of_sale.CustomerFacingQR template:
https://github.com/odoo/odoo/commit/49116abd0916ac835e0ad1f6ba1714b8e80c4272

This commit reintroduces the QR code on the customer display.

closes odoo/odoo#207581

X-original-commit: 2439af09c7586a22822549164b677fa6fb805359
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>